fix delegates retrieval
